+++ game.h
@@ -9,12 +9,12 @@
 
 typedef struct
 {
-  bool   state;
-  int*   boats;
-  Board* board_p1;
-  Board* board_p2;
+  bool     state;
+  uint8_t* boats;
+  Board*   board_p1;
+  Board*   board_p2;
 } Game;
 
-Game* construct_game(uint8_t height, uint8_t width, int* boats);
+Game* construct_game(uint8_t height, uint8_t width, uint8_t* boats);
 void destruct_game(Game* game);
 void print_game(Game* game);
